Added Instant Feedback Mode
Added a little mode that allows for instant feedback upon clicking the button instead of having the next and previous buttons. It acts like a regular quiz but without the previous and next buttons and when you click an answer it will show which is the correct one by making it green and the other questions red.